test: move to RTL
