+++ packages/components/src/ActionBar/ActionBar.test.js
@@ -1,7 +1,58 @@
+/* eslint-disable react/prop-types */
 import React from 'react';
-import { mount } from 'enzyme';
+// rewrite thoses test with react-testing-library
+import { render, screen, configure, within } from '@testing-library/react';
 import ActionBar from './ActionBar.component';
-import { Action } from '../Actions';
+configure({ testIdAttribute: 'data-test' });
+
+const primaryAction = {
+	label: 'Primary',
+	icon: 'fa fa-asterisk',
+	onClick: jest.fn(),
+	bsStyle: 'primary',
+};
+
+const secondaryAction = {
+	label: 'Secondary',
+	icon: 'fa fa-asterisk',
+	onClick: jest.fn(),
+};
+
+const splitAction = {
+	displayMode: 'splitDropdown',
+	label: 'Add File',
+	icon: 'fa fa-plus',
+	onClick: jest.fn(),
+	items: [
+		{
+			label: 'From Local',
+			onClick: jest.fn(),
+		},
+		{
+			label: 'From Remote',
+			onClick: jest.fn(),
+		},
+	],
+	emptyDropdownLabel: 'No option',
+};
+
+const btnGroupAction = {
+	displayMode: 'btnGroup',
+	actions: [
+		{
+			label: 'hidden mean tooltips',
+			icon: 'talend-cog',
+			hideLabel: true,
+			onClick: jest.fn(),
+		},
+		{
+			label: 'you are a super star',
+			icon: 'talend-badge',
+			hideLabel: true,
+			onClick: jest.fn(),
+		},
+	],
+};
 
 describe('ActionBar', () => {
 	it('should trigger onClick callback on Action click', () => {
@@ -16,8 +67,8 @@ describe('ActionBar', () => {
 
 		// when
 		const actionBar = <ActionBar {...props} />;
-		const wrapper = mount(actionBar);
-		wrapper.find(Action).at(0).simulate('click');
+		render(actionBar);
+		screen.getByText('Preparations').click();
 
 		// then
 		expect(onClickMock).toHaveBeenCalled();
@@ -26,7 +77,11 @@ describe('ActionBar', () => {
 	it('should support custom component', () => {
 		// given
 		function MyAction(props) {
-			return <div className="my-custom-action" {...props} />;
+			return (
+				<div className="my-custom-action" {...props}>
+					{props.label}
+				</div>
+			);
 		}
 		const props = {
 			selected: 0,
@@ -43,13 +98,11 @@ describe('ActionBar', () => {
 
 		// when
 		const actionBar = <ActionBar {...props} />;
-		const wrapper = mount(actionBar);
-		const render = wrapper.find('.my-custom-action').first();
+		render(actionBar);
+		const element = screen.getByText('Preparations');
 
 		// then
-		expect(render.hasClass('my-custom-action')).toBe(true);
-		expect(render.name()).toBe('div');
-		expect(render.props()).toMatchSnapshot();
+		expect(element).toHaveClass('my-custom-action');
 	});
 
 	it('should render the number of selected items', () => {
@@ -62,12 +115,9 @@ describe('ActionBar', () => {
 		};
 
 		// when
-		const wrapper = mount(<ActionBar {...props} />);
-		const count = wrapper.find('span.tc-actionbar-selected-count');
-
+		render(<ActionBar {...props} />);
 		// then
-		expect(count.length).toBe(1);
-		expect(count.text()).toEqual(`${props.selected} selected`);
+		expect(screen.getByTestId('selected-count')).toHaveTextContent(`${props.selected} selected`);
 	});
 
 	it('should not render the number of selected items', () => {
@@ -81,10 +131,121 @@ describe('ActionBar', () => {
 		};
 
 		// when
-		const wrapper = mount(<ActionBar {...props} />);
-		const count = wrapper.find('span.tc-actionbar-selected-count');
+		render(<ActionBar {...props} />);
 
 		// then
-		expect(count.length).toBe(0);
+		expect(screen.queryByTestId('selected-count')).not.toBeInTheDocument();
+	});
+
+	it('should render no-selected actions, all on left ', () => {
+		// given
+		const props = {
+			selected: 0,
+			actions: {
+				left: [primaryAction, secondaryAction, splitAction],
+			},
+		};
+		// when
+		render(<ActionBar {...props} />);
+		const left = document.getElementsByClassName('navbar-left')[0];
+		// then
+
+		expect(within(left).getByText('Primary')).toBeInTheDocument();
+		expect(within(left).getByText('Secondary')).toBeInTheDocument();
+		expect(within(left).getByText('Add File')).toBeInTheDocument();
+	});
+	it('should render no-selected actions, some on left, the other on right', () => {
+		// given
+		const props = {
+			selected: 0,
+			actions: {
+				left: [primaryAction],
+				right: [secondaryAction, splitAction],
+			},
+		};
+		// when
+		render(<ActionBar {...props} />);
+		const left = document.getElementsByClassName('navbar-left')[0];
+		const right = document.getElementsByClassName('navbar-right')[0];
+		// then
+		expect(within(left).getByText('Primary')).toBeInTheDocument();
+		expect(within(right).getByText('Secondary')).toBeInTheDocument();
+		expect(within(right).getByText('Add File')).toBeInTheDocument();
+	});
+	it('should render no-selected actions, all on right', () => {
+		// given
+		const props = {
+			selected: 0,
+			actions: {
+				right: [primaryAction, secondaryAction, splitAction],
+			},
+		};
+		// rewrite thoses test with react-testing-library
+		// when
+		render(<ActionBar {...props} />);
+		const right = document.getElementsByClassName('navbar-right')[0];
+		// then
+		expect(within(right).getByText('Primary')).toBeInTheDocument();
+		expect(within(right).getByText('Secondary')).toBeInTheDocument();
+		expect(within(right).getByText('Add File')).toBeInTheDocument();
+	});
+	it('should render selected count and multi-selected actions, all on left', () => {
+		// given
+		const props = {
+			selected: 1,
+			actions: {},
+			multiSelectActions: {
+				left: [secondaryAction, splitAction],
+			},
+		};
+		// when
+		render(<ActionBar {...props} />);
+		const left = document.getElementsByClassName('navbar-left')[0];
+		// then
+		expect(within(left).getByText('1 selected')).toBeInTheDocument();
+		expect(within(left).getByText('Secondary')).toBeInTheDocument();
+		expect(within(left).getByText('Add File')).toBeInTheDocument();
+	});
+	it(
+		'should render selected count and multi-selected actions,' +
+			' count and some actions on left, the other on right',
+		() => {
+			// given
+			const props = {
+				selected: 1,
+				actions: {},
+				multiSelectActions: {
+					left: [primaryAction],
+					right: [secondaryAction, splitAction],
+				},
+			};
+			// when
+			render(<ActionBar {...props} />);
+			// then
+			const left = document.getElementsByClassName('navbar-left')[0];
+			const right = document.getElementsByClassName('navbar-right')[0];
+			expect(within(left).getByText('1 selected')).toBeInTheDocument();
+			expect(within(left).getByText('Primary')).toBeInTheDocument();
+			expect(within(right).getByText('Secondary')).toBeInTheDocument();
+			expect(within(right).getByText('Add File')).toBeInTheDocument();
+		},
+	);
+	it('should render selected count and multi-selected actions, all on right', () => {
+		// given
+		const props = {
+			selected: 1,
+			actions: {},
+			multiSelectActions: {
+				right: [primaryAction, secondaryAction, splitAction],
+			},
+		};
+		// when
+		render(<ActionBar {...props} />);
+		// then
+		const right = document.getElementsByClassName('navbar-right')[0];
+		expect(screen.getByText('1 selected')).toBeInTheDocument();
+		expect(within(right).getByText('Primary')).toBeInTheDocument();
+		expect(within(right).getByText('Secondary')).toBeInTheDocument();
+		expect(within(right).getByText('Add File')).toBeInTheDocument();
 	});
 });
